Stop rsyncing seconds files for each cycle.
The purpose of the files was to record the exact number of seconds
elapsed since the benchmark started.  Due to David's previous
improvements to the snapshotting and rsyncing process, snapshots have
been occurring at exactly 20 second intervals without fail.  So we can
now confidently infer the seconds elapsed with just the snapshot number.