Determine openssl-1.0.1f cert file paths at runtime.
The A/B testing framework executes benchmarks on different VMs than they
were built on.  For the openssl-1.0.1f benchmark, this means that the
path to the server.pem and server.key files is different at runtime than
it was at compile-time.  This commit copies the keys to the runner VM
and adds functionality to select the appropriate path at runtime.