Use prebuilt Docker image for dispatcher.
Reduces startup time from 2m49s to 1m57s.  Pulls the latest image built
at gcr.io/fuzzer-test-suite/dispatcher, which is built from
base-image/Dockerfile-dispatcher and contains all benchmark dependencies
and a recent clang (as of this commit).

We will need to rebuild that image whenever we want to use a newer clang
on the dispatcher.  We may be able to automate those builds using Google
Container Builder if we really care about using ToT clang.