Extract Clang revision from image tag when clang --version fails.
Clang nightly builds from apt.llvm.org no longer state the revision of
the build in clang --version.  For now, we can add the revision as a
tag to dispatcher images in GCR.  Then if clang --version fails, we can
read the Clang revision from the tag on the latest dispatcher image.